refactor: confirm_delete系を common.py から prompt_util.py に移動
