+++ src/redi/cli/common.py
@@ -2,43 +2,12 @@ import os
 import subprocess
 import tempfile
 
-from prompt_toolkit import Application, prompt
+from prompt_toolkit import Application
 from prompt_toolkit.key_binding import KeyBindings
 from prompt_toolkit.layout import HSplit, Layout, Window
 from prompt_toolkit.layout.controls import FormattedTextControl
 
 from redi.config import editor
-from redi.i18n import messages
-
-
-def confirm_delete(summary: str) -> None:
-    print(summary)
-    try:
-        confirm = prompt(messages.prompt_confirm_delete).strip().lower()
-    except (KeyboardInterrupt, EOFError):
-        print(messages.canceled)
-        exit(1)
-    if confirm != "yes":
-        print(messages.canceled)
-        exit(1)
-
-
-def confirm_delete_with_identifier(
-    summary: str, expected: str, field_label: str
-) -> None:
-    print(summary)
-    try:
-        entered = prompt(
-            messages.prompt_confirm_delete_with_identifier.format(
-                label=field_label, expected=expected
-            )
-        ).strip()
-    except (KeyboardInterrupt, EOFError):
-        print(messages.canceled)
-        exit(1)
-    if entered != expected:
-        print(messages.canceled_field_mismatch.format(field=field_label))
-        exit(1)
 
 
 SUBCOMMAND_ALIASES: dict[str, str] = {
