+++ src/redi/cli/prompt_util.py
@@ -1,6 +1,7 @@
 import re
 from datetime import date
 
+from prompt_toolkit import prompt
 from prompt_toolkit.document import Document
 from prompt_toolkit.key_binding import KeyBindings
 from prompt_toolkit.keys import Keys
@@ -132,3 +133,33 @@ def digit_and_period_key_bindings() -> KeyBindings:
             event.current_buffer.insert_text(event.data)
 
     return kb
+
+
+def confirm_delete(summary: str) -> None:
+    print(summary)
+    try:
+        confirm = prompt(messages.prompt_confirm_delete).strip().lower()
+    except (KeyboardInterrupt, EOFError):
+        print(messages.canceled)
+        exit(1)
+    if confirm != "yes":
+        print(messages.canceled)
+        exit(1)
+
+
+def confirm_delete_with_identifier(
+    summary: str, expected: str, field_label: str
+) -> None:
+    print(summary)
+    try:
+        entered = prompt(
+            messages.prompt_confirm_delete_with_identifier.format(
+                label=field_label, expected=expected
+            )
+        ).strip()
+    except (KeyboardInterrupt, EOFError):
+        print(messages.canceled)
+        exit(1)
+    if entered != expected:
+        print(messages.canceled_field_mismatch.format(field=field_label))
+        exit(1)
