+++ src/redi/cli/time_entry_command.py
@@ -5,13 +5,13 @@ from prompt_toolkit.completion import WordCompleter
 from prompt_toolkit.validation import Validator
 
 from redi.cli.common import (
-    confirm_delete,
     inline_checkbox,
     inline_choice,
     resolve_alias,
 )
 from redi.cli.prompt_util import (
     HourValidator,
+    confirm_delete,
     digit_and_period_key_bindings,
     digit_only_key_bindings,
 )
